Agregado nuevo nombre Hero-component
